refactor(server): move thread bootstrap out of the WebSocket handler
The create-and-start sequence behind thread.turn.start with a bootstrap
lived inside the per-connection RPC closure in ws.ts. The MCP server
needs the same sequence, so it now lives in a ThreadBootstrap service.
No behavior change: same commands, fences, cleanup, and setup activity.
